+++ src/main/java/prgms/vouchermanagementapp/storage/Customers.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
-import prgms.vouchermanagementapp.customer.Customer;
+import prgms.vouchermanagementapp.domain.Customer;
 
 import javax.sql.DataSource;
 import java.util.Map;
