+++ src/main/java/prgms/vouchermanagementapp/storage/MemoryVouchers.java
@@ -2,9 +2,9 @@ package prgms.vouchermanagementapp.storage;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+import prgms.vouchermanagementapp.domain.Voucher;
 import prgms.vouchermanagementapp.storage.model.MemoryVoucherRecord;
 import prgms.vouchermanagementapp.storage.model.VoucherRecord;
-import prgms.vouchermanagementapp.voucher.model.Voucher;
 
 import java.util.ArrayList;
 import java.util.HashMap;
