+++ src/main/java/prgms/vouchermanagementapp/storage/Vouchers.java
@@ -1,7 +1,7 @@
 package prgms.vouchermanagementapp.storage;
 
+import prgms.vouchermanagementapp.domain.Voucher;
 import prgms.vouchermanagementapp.storage.model.VoucherRecord;
-import prgms.vouchermanagementapp.voucher.model.Voucher;
 
 public interface Vouchers {
