+++ src/main/java/prgms/vouchermanagementapp/voucher/VoucherManager.java
@@ -1,11 +1,11 @@
 package prgms.vouchermanagementapp.voucher;
 
 import org.springframework.stereotype.Component;
-import prgms.vouchermanagementapp.io.model.Amount;
-import prgms.vouchermanagementapp.io.model.Ratio;
+import prgms.vouchermanagementapp.domain.Voucher;
+import prgms.vouchermanagementapp.domain.model.Amount;
+import prgms.vouchermanagementapp.domain.model.Ratio;
 import prgms.vouchermanagementapp.storage.Vouchers;
 import prgms.vouchermanagementapp.storage.model.VoucherRecord;
-import prgms.vouchermanagementapp.voucher.model.Voucher;
 
 @Component
 public class VoucherManager {
